[FixedRewardGridWorld] Use float for rewards

+++ rlpy/domains/fixed_reward_grid_world.py
@@ -20,13 +20,13 @@ class FixedRewardGridWorld(GridWorld):
     )
 
     def _load_map(self, mapfile):
-        map_and_reward = np.loadtxt(mapfile, dtype=np.int32)
+        map_and_reward = np.loadtxt(mapfile, dtype=np.float64)
         mshape = map_and_reward.shape
         if mshape[1] * 2 != mshape[0]:
             raise ValueError("Invalid map with shape {}".format(mshape))
         col = mshape[0] // 2
         self.reward_map = map_and_reward[col:]
-        return map_and_reward[:col]
+        return map_and_reward[:col].astype(np.int32)
 
     def __init__(
         self,
